teste_4/backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
pandas==2.0.3
python-dotenv==1.0.0
gunicorn==20.1.0